+++ src/navigation/Stack.js
@@ -13,6 +13,7 @@ import {
   Splash,
   TestingComponent,
   Settings,
+  ChangePassword
 } from '../screens';
 import MainApp from './MainApp';
 
@@ -40,6 +41,7 @@ function Navigation() {
       <Stack.Screen name="SearchProductScreen" component={SearchProduct} />
       <Stack.Screen name="TestingComponentScreen" component={TestingComponent} />
       <Stack.Screen name="SettingsScreen" component={Settings} />
+      <Stack.Screen name="ChangePasswordScreen" component={ChangePassword} />
     </Stack.Navigator>
   );
 }
