+++ src/screens/Pengaturan/styles.js
@@ -1,10 +1,8 @@
 import {StyleSheet} from 'react-native';
-import {SIZES} from '../../themes';
 
 export default StyleSheet.create({
-  form: {
-    marginHorizontal: SIZES.padding,
-    marginTop: 24,
-    marginBottom: 30,
+  menuWrapper: {
+    marginTop: 20,
+    paddingHorizontal: 20,
   },
 });
