+++ src/screens/UbahPassword/styles.js
@@ -1,0 +1,10 @@
+import {StyleSheet} from 'react-native';
+import {SIZES} from '../../themes';
+
+export default StyleSheet.create({
+  form: {
+    marginHorizontal: SIZES.padding,
+    marginTop: 24,
+    marginBottom: 30,
+  },
+});
