+++ src/screens/index.js
@@ -15,9 +15,11 @@ import SearchProduct from './SearchProduct';
 import DetailProductSeller from './DetailProductSeller';
 import TestingComponent from './TestingComponent';
 import Settings from './Pengaturan';
+import ChangePassword from './UbahPassword';
 
 export {
   Settings,
+  ChangePassword,
   DaftarJual,
   DetailProduct,
   Account,
